# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "description": "Возвращает HTML страницу со списком всех метрик (gauge и counter) из базы данных.",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Получение всех метрик в HTML формате",
                "responses": {
                    "200": {
                        "description": "HTML страница со списком метрик",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Проверяет доступность подключения к бд",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Проверка связи от базой данных(далее бд)",
                "responses": {
                    "200": {
                        "description": "ОК",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка соединения с бд",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update": {
            "post": {
                "description": "Обновляет метрику. Поддерживает два формата:",
                "consumes": [
                    "application/json",
                    "text/plain"
                ],
                "produces": [
                    "application/json",
                    "text/plain"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Обновление метрики",
                "parameters": [
                    {
                        "description": "Метрика в JSON формате",
                        "name": "metric",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Metrics"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Строка со статусом OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос: bad gauge value, bad counter value, unknown metric type, metric ID is required, gauge value is required, counter delta is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера: store error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update/": {
            "post": {
                "description": "Обновляет метрику. Поддерживает два формата:",
                "consumes": [
                    "application/json",
                    "text/plain"
                ],
                "produces": [
                    "application/json",
                    "text/plain"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Обновление метрики",
                "parameters": [
                    {
                        "description": "Метрика в JSON формате",
                        "name": "metric",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Metrics"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Строка со статусом OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос: bad gauge value, bad counter value, unknown metric type, metric ID is required, gauge value is required, counter delta is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера: store error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update/{type}/{name}/{value}": {
            "post": {
                "description": "Обновляет метрику. Поддерживает два формата:",
                "consumes": [
                    "application/json",
                    "text/plain"
                ],
                "produces": [
                    "application/json",
                    "text/plain"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Обновление метрики",
                "parameters": [
                    {
                        "description": "Метрика в JSON формате",
                        "name": "metric",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Metrics"
                        }
                    },
                    {
                        "enum": [
                            "gauge",
                            "counter"
                        ],
                        "type": "string",
                        "description": "Тип метрики",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Имя метрики",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Значение метрики",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Строка со статусом OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос: bad gauge value, bad counter value, unknown metric type, metric ID is required, gauge value is required, counter delta is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера: store error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/updates": {
            "post": {
                "description": "Принимает массив метрик в JSON формате и обновляет их все за один запрос",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Пакетное обновление метрик",
                "parameters": [
                    {
                        "description": "Массив метрик для обновления",
                        "name": "metrics",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Metrics"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пример: {\\\"status\\\":\\\"OK\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Пример: {\\\"error\\\":\\\"validation failed\\\",\\\"details\\\":[\\\"metric[0]: ID is required\\\"]}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Пример: {\\\"error\\\":\\\"failed to update metric Alloc\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/value": {
            "post": {
                "description": "Принимает JSON с полями id и type, возвращает структуру метрики с полями ID, Type, Value.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Получение метрики в JSON формате",
                "parameters": [
                    {
                        "description": "Метрика для поиска (нужны только id и type)",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Metrics"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Найденная метрика",
                        "schema": {
                            "$ref": "#/definitions/model.Metrics"
                        }
                    },
                    "400": {
                        "description": "Неверный формат JSON, отсутствуют обязательные поля, неизвестный тип метрики",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Метрика не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/value/{type}/{name}": {
            "get": {
                "description": "Получает значение метрики по её типу и имени.",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Получение значения метрики",
                "parameters": [
                    {
                        "enum": [
                            "gauge",
                            "counter"
                        ],
                        "type": "string",
                        "description": "Тип метрики",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Имя метрики",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Значение метрики в виде строки",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный тип метрики",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Метрика не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Metrics": {
            "type": "object",
            "properties": {
                "delta": {
                    "description": "целочисленное поле, используется для хранения значения метрики типа counter",
                    "type": "integer"
                },
                "hash": {
                    "description": "контрольная сумма процерки целостности данных",
                    "type": "string"
                },
                "id": {
                    "description": "уникальный индетификатор",
                    "type": "string"
                },
                "type": {
                    "description": "тим метрики",
                    "type": "string"
                },
                "value": {
                    "description": "число с плавающей точкой, используется для хранения значения метрики типа gauge",
                    "type": "number"
                }
            }
        }
    }
}